+++ tsconfig.json
@@ -16,5 +16,8 @@
   "include": [
     "src/**/*",
     "tests/**/*"
+  ],
+  "exclude": [
+    "node_modules/**/*"
   ]
 }
